+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_giphy/giphy_ui.dart';
+import 'package:flutter_giphy_picker/giphy_ui.dart';
 
 void main() {
   runApp(const MyApp());