+++ lib/ui/View/Gifs/gif_tab.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_giphy/giphy_api.dart';
-import 'package:flutter_giphy/ui/View/Gifs/Favorites/favorites.dart';
-import 'package:flutter_giphy/ui/View/Gifs/gifs.dart';
-import 'package:flutter_giphy/ui/View/Gifs/translate.dart';
-import 'package:flutter_giphy/ui/giphy_config.dart';
-import 'package:flutter_giphy/ui/giphy_result.dart';
+import 'package:flutter_giphy_picker/giphy_api.dart';
+import 'package:flutter_giphy_picker/ui/View/Gifs/Favorites/favorites.dart';
+import 'package:flutter_giphy_picker/ui/View/Gifs/gifs.dart';
+import 'package:flutter_giphy_picker/ui/View/Gifs/translate.dart';
+import 'package:flutter_giphy_picker/ui/giphy_config.dart';
+import 'package:flutter_giphy_picker/ui/giphy_result.dart';
 
 class GifTabView extends StatefulWidget {
   final GiphyAPI giphyAPI;